Revert "Rework update systems + Change names of namespaces"
This reverts commit 1a0aa429f225e22acf9700bc6a25884c8dfe0ca4.
