--- ProcgenGame/Core/Components/InputComponent.cs
@@ -1,6 +1,0 @@
-﻿namespace ProcgenGame.Core.Components;
-
-sealed class InputComponent : Component
-{
-    public PlayerInputs InputFlags { get; set; }
-}
