+++ ProcgenGame/Core/Entities/Entity.cs
@@ -6,10 +6,10 @@ class Entity
 {
     readonly int _id = GlobalState.AutoId();
     readonly List<Component> _components = new();
-    readonly EntityRegistry _entityRegister;
-    readonly ComponentRegistry _componentRegister;
+    readonly EntityRegister _entityRegister;
+    readonly ComponentRegister _componentRegister;
 
-    internal Entity(EntityRegistry entityRegister, ComponentRegistry componentRegister)
+    internal Entity(EntityRegister entityRegister, ComponentRegister componentRegister)
     {
         _entityRegister = entityRegister;
         _componentRegister = componentRegister;
