+++ ProcgenGame/Core/Entities/EntitySpawner.cs
@@ -4,23 +4,22 @@ namespace ProcgenGame.Core.Entities;
 
 sealed class EntitySpawner
 {
-    readonly EntityRegistry _entityRegistry;
-    readonly ComponentRegistry _componentRegistry;
+    readonly EntityRegister _entityRegister;
+    readonly ComponentRegister _componentRegister;
 
-    internal EntitySpawner(EntityRegistry entityRegistry, ComponentRegistry componentRegistry)
+    internal EntitySpawner(EntityRegister entityRegister, ComponentRegister componentRegister)
     {
-        _entityRegistry = entityRegistry;
-        _componentRegistry = componentRegistry;
+        _entityRegister = entityRegister;
+        _componentRegister = componentRegister;
     }
 
     public Entity SpawnPlayer(Vector2 position = new Vector2())
     {
-        var entity = new Entity(_entityRegistry, _componentRegistry);
+        var entity = new Entity(_entityRegister, _componentRegister);
 
         entity.AddComponent(new TransformComponent() { Position = position, Size = 32 });
         entity.AddComponent(new PhysicsComponent() { Speed = 100 });
-        entity.AddComponent(new CollisionComponent() { IsSolid = true });
-        entity.AddComponent(new InputComponent());
+        entity.AddComponent(new CollisionComponent() { IsSolid = true }); 
         entity.AddComponent(new DrawComponent() { Color = Color.Red });
 
         return entity;
